Finished merging, still needs testing

+++ src/hubo-read.c
@@ -210,7 +210,7 @@ void huboLoop() {
 		int jnt = 0;
 
 		clock_gettime( 0,&t2);
-		printf("t = %i.%09i\n",t2.tv_sec,t2.tv_nsec);
+		printf("t = %i.%09i\n",(int)t2.tv_sec,(int)t2.tv_nsec);
 		for( i = 0; i < HUBO_JOINT_COUNT; i++) {
 			jnt = i;
 			if(H_param.joint[jnt].name[0] != 0){
